// resources/js/models/UserModel.ts
export type UserModel = {
  id: string
  name: string
  email: string
  legal_document: any
  phone: any
  cellphone: any
  address: any
  street: any
  city: any
  state: any
  zipcode: any
  country: any
  role: string
  status: string
  email_verified_at: any
  two_factor_confirmed_at: any
  current_team_id: any
  profile_photo_path: any
  created_at: string
  updated_at: string
  profile_photo_url: string
  two_factor_enabled: boolean
}